"rmc make" uses "spock-filter" and "blame-warnings"
The spock-filter controls how signals are sent to build commands with
filtered output. The blame-warnings filter reports who is responsible
for each compiler warning by using the Git history. The blame-warnings
is used only if it's available in $PATH and it works (it's distributed
from a different repository and thus might not always be installed).