Assertion enable/disable is separated from debug and optimize settings
The rmc_assertions setting now takes one or two arguments, which can
appear in any order.

One argument is the word 'enabled' or 'disabled' to control whether
optional assertions are compiled or not.

The other argument is the word 'abort', 'throw', 'exit', or
'ambivalent' and indicates how compiled-in assertions (required
assertions and enabled optional assertions) behave when they fail if
the behavior is not specified at runtime.